ffi: sanity check on object in ffi_val_get.
* ffi.c (ffi_val_get): Throw an exception if the object
doesn't appear valid.

* gc.c (valid_object_p): New function.  Invalid objects are
those that are pointers, but either not into a heap, or else
to a freed object or to an object that is marked by the
garbage collector (should only be seen while GC is running).
All others are valid.  There can be false positives here: a
value with the LIT tag is in fact a pointer, but we don't
check whether that is valid.

* gc.c (valid_object_p): Declared.

* txr.1: Remarks added to documentation of FFI val type.