ffi: buffix: alignment still wrong for ptr types.
For pointers, the code still reflects the wrong assumption
that the size and alignment are the same.

* ffi.c (make_ffi_type_pointer): Remove the size argument;
it's always passed as sizeof (mem_t *), which is pointless.
Just initialize the size and align members in a fixed way.
Use sizeof and alignof to do this right.
(ffi_type_compile): Remove size argument from all
make_ffi_type_pointer calls.