ffi: evaluate expressions in type notation.
Places where an integer constant was previously allowed
now allow an expression.

The way enum syntax works is now different. A temporary
lexical environment is created, and each enumeration is
bound in that environment. The value expressions are
evaluated in that environment. The decision to allow
keyword symbols to be enumeration contants is retracted.

* eval.h (special_s): Declared.

* ffi.c (ffi_eval_expr): New static function.
(make_ffi_type_enum): Enums are introduced into a temporary
environment, in which the value expressions are evaluated. By
this means, the expressions can refer can refer to previous
enums and do arbitrary computation. Also, we drop the
requirement that enums can be keyword symbols.
(ffi_type_compile): Array dimension and buf size are evaluated
as expresions.  Array and buffer syntax is transformed by
substitution of the evaluated size.

* txr.1: Documented use of expressions in FFI type notation.