ffi: support sel operation on carray.
Thus (select ca '(0 3 4 ...)) works and so does the
sytnax [ca '(0 3 4 ...)]. This is inefficiently implemented.
The selected elements are extracted to a list which is then
converted to a carray of the same kind agan.

* ffi.c (carray_list): New function.
(ffi_init): Register carray-list intrinsic.

* ffi.h (carray_list): Declared.

* lib.c (make_like): Add carray case, so we can turn
a list into a carray based on an example carray.
This uses carray_list, with the type pulled from the
original carray. The target isn't null terminated.
(sel): Handle carray via vector case.

* txr.1: Document changes in select and make-like.