bugfix: autoload syntactic places.
TXR Lisp doesn't autoload the definition of places.
For instance if a (set (foo x) y) appears out of the
blue and foo is keyed for autoload, it doesn't happen.

The right place to fix this is to do the autload check
in the place macro expander, since every place form is
tried as a macro.

We need to expose the lisplib_try_load function as a Lisp
intrinsic.

* lisplib.c (lisplib_init): Register sys:try-load
intrinsic, mapped to lisplib_try_load.

* share/txr/stdlib/place.tl (sys:get-place-macro):
New function.
(sys:pl-expand): Use sys:get-place-macro instead of
direct lookup in *place-macro* hash. The new function
tries to autoload the symbol if it finds no place
macro for it.