New length method on structures.
With this change, nreverse is possible on a struct,
if it has lambda, lambda-set and length methods.

* lib.c (length_s): New symbol variable.
(length): Check for length method and call it,
in preference to checking for car method and falling
back on cdr-based length.
(empty): If object has a length method, call it and
compare result to zero.
(obj_init): Initialize length_s.

* txr.1: Documented. We now have the concepts of structs
being vector-like sequences or list-like sequences.