linenoise: replace linefeeds with CR in previous result.
This fixes the issue that when a multi-line result is pasted
into the edit buffer with Ctrl-X P, the line breaks appear in
it as linefeeds, displayed as ^J. In linenoise, we need line
breaks to be carriage returns.

* linenoise/linenoise.c (lino_set_result): replace newlines
with carriage returns in the given string.

* linenoise/linenoise.h (lino_set_result): Add comment that
function takes ownership of memory assumed to be malloced,
and that it modifies it.