format: support leading zero printing for floats.
For a floating-point argument, the following behavior
is supported by the ~a, ~s and ~f format directives:
if the precision has a leading zero, then leading zeros
are added, up to one less than the field width.

* stream.c (formatv): Set the precision to one less than
the field width when rendering floats, rather than to
zero. The output_num code takes the precision as the cue
for padding zeros.

* txr.1: Updated format documentation, and clarified
a few things also.