find, pos: optimize and support objects properly.
* lib.c (find, pos): Provide specialized behavior based
on type and test and key functions. Lists and list-like
objects are treated by marching down with cdr.  Vectors are
traversed with numeric index, as are vector-like objects which
exhibit a length function.  A special optimization is put in
for non-lazy strings which use identity as their key function,
and one of the built-in equality operators for the test
function: wcschr is used on the underlying C string.