expander: do dot-to-apply for meta-expressions.
The dot-to-apply transformation is now applied when
meta-expressions like @foo and @(bar) apparently occur in the
dot position.

This change is made in anticipation of a rewrite of the op
macro, in which the @1, @2,  and @rest arguments will be
implemented as macrolets, rather than the ad-hoc, hacky code
walk currently performed in the transform_op function.

* eval.c (dot_meta_list_p): New static function.
(dot_to_apply): Detect the presence of a sys:var or
sys:expr argument in a form. If found, then turn
it and the remaining forms into a single compound
form which replaces them.

* txr.1: Update doc under Dot Position in Function Calls.