txr -i honored despite parse-time exception.
If an error is thrown while parsing a .txr file
or while reading and evaluating the forms of a .tl
file.

* parser.y (parse_once, parse): Wording change in message when
exception is caught. Only exceptions derived from error are
caught.

* txr.c (parse_once_noerr, read_eval_stream_noerr): New static
functions.
(txr_main): Use parse_once_noerr and read_eval_stream_noerr
instead of parse_once and read_eval_stream.
Don't exit if a TXR file has parser errors; in that situation,
exit only if interactive mode is not requested, otherwise
go interactive. Make sure *self-path* is registered to the
name of the input source in this case also.

* unwind.h (ignerr_func_body): New macro.