linenoise: bugfix: cancel extended mode on Enter.
* linenoise/linenoise.c (edit): If Enter is processed while in
Ctrl-X extended command mode, that mode must be explicitly
canceled by resetting the extended local flag. Not doing this
became an issue when the Enter callback mechanism was
introduced to detect incomplete lines. At that point, entering
Ctrl-X Enter on an incomplete line caused linenoise
to insert a newline, but stay in extended mode.