regex: bugfix: incorrect use of nfa_move_closure.
This goes back to November 2, 2009 commit,
"Start of implementation for freestyle matching",
6191fbb2ca7a9ac339dd3994bdea8273ceb0a24d
It is exposed if we perform an epsilon threading
optimization on the NFA graph.

* regex.c (regex_machine_feed): We must pass the new,
incremented value of the visited counter, to nfa_move_closure.
States have already been tagged with the old value before
the call to regex_machine_feed, so we risk failing to
visit some states and include them in the closure.