regex: bug printing #/abc(def|ghi)/
This was broken by the July 16 commit
"regex: don't print superfluous parens around classes",
2411f779f47c441659720ad0ddcabf91df1d2529.

* regex.c (print_rec): If an (or ...) appears as
a compound element, it must be rendered in parentheses;
or_s must be handled here just like and_s.  Prior to the
faulty commit, this was implicitly true because the logic was
inverted and wasn't ruling out or_s.