places: use Lisp-1 macroexpansion where needed.
A test case for this very subtle bug is this:

  (let ((v (list 1 2 3)))
    (symacrolet ((x v))
      (flet ((x () 42))
        (set [x 0] 0))))

Because x is being evaluated in the DWIM brackets which
flatten the two namespaces into one, it must be treated as a
reference to the flet, and so [x 0] denotes the function call.
The assignment is erroneous.

The incorrect behavior being fixed is that the places code
macro-expands x in the Lisp-2 style under which the symacrolet
is not shadowed by the flet. The substitution of v takes
place, and the assignment assigns to [v 0].

* share/txr/stdlib/place.tl (sys:l1-setq, sys:l1-val):
Use macroexpand-lisp1 rather than macroexpand.